Vladimir Buell <farjumper@mail.ru>:
I've tried to connect to my corporate VPN, but the only message I got was:

vpnc: xauth packet unsupported: ATTRIBUTES_NOT_SUPPORTED

Problem was found in xauth_type. In my case request attribute "xauth_type" consisted 5
+as value. There is quick
fix for this issue below. But...

In the chapter 6.2 of the draft-beaulieu-ike-xauth-02.txt we can read:

"XAUTH-TYPE - The type of extended authentication requested whose
   values are described in the next section.  This is an optional
   attribute for the ISAKMP_CFG_REQUEST and ISAKMP_CFG_REPLY messages.
   If the XAUTH-TYPE is not present, then it is assumed to be Generic.
   The XAUTH-TYPE in a REPLY MUST be identical to the XAUTH-TYPE in the
   REQUEST."

So I think the better way is to send reply with the same xauth_type value as in
request. But I can't test it because our server always use 5 and never 0.