Jukka Salmi <j+vpnc@2009.salmi.ch>
the Perl scripts makeman.pl, enum2debug.pl and pcf2vpnc all expect perl
to be available in /usr/bin.  Of course this causes problems if perl is
installed somewhere else (like, e.g., on systems using pkgsrc, where
perl is in /usr/pkg/bin).

What about the attached patch?  It assumes env(1) to be in /usr/bin (and
perl to be in the $PATH), but that's much more likely...